% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{auction}
\alias{auction}
\title{auction}
\usage{
auction(x, y)
}
\arguments{
\item{x}{= internal representation of a bridge hand}

\item{y}{= internal representation of a bridge hand}
}
\description{
full auction assuming no competition
}
\keyword{auction}
