% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{cheapest}
\alias{cheapest}
\title{cheapest}
\usage{
cheapest(x, y)
}
\arguments{
\item{x}{= existing bid}

\item{y}{= new strain}
}
\description{
cheapest bid in strain <y> after bid <x>
}
\keyword{cheapest}
