% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{example.hands}
\alias{example.hands}
\title{example.hands}
\usage{
example.hands(fcn, fun)
}
\arguments{
\item{fcn}{= function that determines when to stop}

\item{fun}{= type of hand you want to focus on}
}
\description{
hand examples
}
\keyword{example.hands}
