% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{example}
\alias{example}
\title{example}
\usage{
example(fcn, fun)
}
\arguments{
\item{fcn}{= function that determines when to stop}

\item{fun}{= type of hand you want to focus on}
}
\description{
hand examples
}
\keyword{example}
