% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{fcn.all.roxygenize}
\alias{fcn.all.roxygenize}
\title{fcn.all.roxygenize}
\usage{
fcn.all.roxygenize(x)
}
\arguments{
\item{x}{= the output file}
}
\description{
roxygenizes all functions
}
\seealso{
Other fcn: \code{\link{fcn.comments.parse}},
  \code{\link{fcn.extract.args}},
  \code{\link{fcn.extract.out}}, \code{\link{fcn.list}},
  \code{\link{fcn.order}}, \code{\link{fcn.path}},
  \code{\link{fcn.roxygenize}},
  \code{\link{fcn.to.comments}}, \code{\link{fcn.to.txt}}
}
\keyword{fcn.all.roxygenize}
