% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{fcn.comments.parse}
\alias{fcn.comments.parse}
\title{fcn.comments.parse}
\usage{
fcn.comments.parse(x)
}
\arguments{
\item{x}{= a string vector (comments section of a function)}
}
\description{
extracts information from the comments
}
\seealso{
Other fcn: \code{\link{fcn.all.roxygenize}},
  \code{\link{fcn.extract.args}},
  \code{\link{fcn.extract.out}}, \code{\link{fcn.list}},
  \code{\link{fcn.order}}, \code{\link{fcn.path}},
  \code{\link{fcn.roxygenize}},
  \code{\link{fcn.to.comments}}, \code{\link{fcn.to.txt}}
}
\keyword{fcn.comments.parse}
