% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{fcn.extract.args}
\alias{fcn.extract.args}
\title{fcn.extract.args}
\usage{
fcn.extract.args(x)
}
\arguments{
\item{x}{= a string vector (argument section of comments)}
}
\description{
vector of arguments with explanations
}
\seealso{
Other fcn: \code{\link{fcn.all.roxygenize}},
  \code{\link{fcn.comments.parse}},
  \code{\link{fcn.extract.out}}, \code{\link{fcn.list}},
  \code{\link{fcn.order}}, \code{\link{fcn.path}},
  \code{\link{fcn.roxygenize}},
  \code{\link{fcn.to.comments}}, \code{\link{fcn.to.txt}}
}
\keyword{fcn.extract.args}
