% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{fcn.list}
\alias{fcn.list}
\title{fcn.list}
\usage{
fcn.list(x = "*")
}
\arguments{
\item{x}{= a string (regular expression)}
}
\description{
Returns the names of objects that are or are not functions
}
\seealso{
Other fcn: \code{\link{fcn.all.roxygenize}},
  \code{\link{fcn.comments.parse}},
  \code{\link{fcn.extract.args}},
  \code{\link{fcn.extract.out}}, \code{\link{fcn.order}},
  \code{\link{fcn.path}}, \code{\link{fcn.roxygenize}},
  \code{\link{fcn.to.comments}}, \code{\link{fcn.to.txt}}
}
\keyword{fcn.list}
