% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{fcn.order}
\alias{fcn.order}
\title{fcn.order}
\usage{
fcn.order()
}
\description{
functions in alphabetical order
}
\seealso{
Other fcn: \code{\link{fcn.all.roxygenize}},
  \code{\link{fcn.comments.parse}},
  \code{\link{fcn.extract.args}},
  \code{\link{fcn.extract.out}}, \code{\link{fcn.list}},
  \code{\link{fcn.path}}, \code{\link{fcn.roxygenize}},
  \code{\link{fcn.to.comments}}, \code{\link{fcn.to.txt}}
}
\keyword{fcn.order}
