% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{fcn.path}
\alias{fcn.path}
\title{fcn.path}
\usage{
fcn.path()
}
\description{
path to function source file
}
\seealso{
Other fcn: \code{\link{fcn.all.roxygenize}},
  \code{\link{fcn.comments.parse}},
  \code{\link{fcn.extract.args}},
  \code{\link{fcn.extract.out}}, \code{\link{fcn.list}},
  \code{\link{fcn.order}}, \code{\link{fcn.roxygenize}},
  \code{\link{fcn.to.comments}}, \code{\link{fcn.to.txt}}
}
\keyword{fcn.path}
