% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{fcn.roxygenize}
\alias{fcn.roxygenize}
\title{fcn.roxygenize}
\usage{
fcn.roxygenize(x, y, n)
}
\arguments{
\item{x}{= a string (function name)}

\item{y}{= a string (library to import)}

\item{n}{= vector of function families}
}
\description{
roxygenized function format
}
\seealso{
Other fcn: \code{\link{fcn.all.roxygenize}},
  \code{\link{fcn.comments.parse}},
  \code{\link{fcn.extract.args}},
  \code{\link{fcn.extract.out}}, \code{\link{fcn.list}},
  \code{\link{fcn.order}}, \code{\link{fcn.path}},
  \code{\link{fcn.to.comments}}, \code{\link{fcn.to.txt}}
}
\keyword{fcn.roxygenize}
