% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{fcn.to.comments}
\alias{fcn.to.comments}
\title{fcn.to.comments}
\usage{
fcn.to.comments(x)
}
\arguments{
\item{x}{= a string (function name)}
}
\description{
returns the comment section
}
\seealso{
Other fcn: \code{\link{fcn.all.roxygenize}},
  \code{\link{fcn.comments.parse}},
  \code{\link{fcn.extract.args}},
  \code{\link{fcn.extract.out}}, \code{\link{fcn.list}},
  \code{\link{fcn.order}}, \code{\link{fcn.path}},
  \code{\link{fcn.roxygenize}}, \code{\link{fcn.to.txt}}
}
\keyword{fcn.to.comments}
