% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{fcn.to.txt}
\alias{fcn.to.txt}
\title{fcn.to.txt}
\usage{
fcn.to.txt(x, y = F, n = F)
}
\arguments{
\item{x}{= a string (function name)}

\item{y}{= a boolean (do/don't return comments)}

\item{n}{= a boolean (return string/vector)}
}
\description{
represents <x> as a string or string vector
}
\seealso{
Other fcn: \code{\link{fcn.all.roxygenize}},
  \code{\link{fcn.comments.parse}},
  \code{\link{fcn.extract.args}},
  \code{\link{fcn.extract.out}}, \code{\link{fcn.list}},
  \code{\link{fcn.order}}, \code{\link{fcn.path}},
  \code{\link{fcn.roxygenize}},
  \code{\link{fcn.to.comments}}
}
\keyword{fcn.to.txt}
