% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{good.suit}
\alias{good.suit}
\title{good.suit}
\usage{
good.suit(x, y = T)
}
\arguments{
\item{x}{= internal representation of a bridge hand}

\item{y}{= T/F depending on whether both criteria are used}
}
\description{
T/F depending on whether suit is good
}
\keyword{good.suit}
