% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{hand.random}
\alias{hand.random}
\title{hand.random}
\usage{
hand.random()
}
\description{
internal representation of a bridge hand
}
\seealso{
Other hand: \code{\link{hand.read}},
  \code{\link{hand.show}}
}
\keyword{hand.random}
