% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{hand.read}
\alias{hand.read}
\title{hand.read}
\usage{
hand.read(x)
}
\arguments{
\item{x}{= a string}
}
\description{
internal representation of a bridge hand
}
\seealso{
Other hand: \code{\link{hand.random}},
  \code{\link{hand.show}}
}
\keyword{hand.read}
