% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{hand.show}
\alias{hand.show}
\title{hand.show}
\usage{
hand.show(x)
}
\arguments{
\item{x}{= internal representation of a bridge hand}
}
\description{
a string
}
\seealso{
Other hand: \code{\link{hand.random}},
  \code{\link{hand.read}}
}
\keyword{hand.show}
