% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{level}
\alias{level}
\title{level}
\usage{
level(x, y)
}
\arguments{
\item{x}{= an integer (point count)}

\item{y}{= an integer vector (level thresholds)}
}
\description{
responder's response to 1m
}
\keyword{level}
