% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{opener.1N}
\alias{opener.1N}
\title{opener.1N}
\usage{
opener.1N(x)
}
\arguments{
\item{x}{= internal representation of a bridge hand}
}
\description{
opener's subsequent bids after 1N
}
\seealso{
Other opener: \code{\link{opener.1Nresp}},
  \code{\link{opener.1Xresp}}, \code{\link{opener.1major}},
  \code{\link{opener.1minor}}, \code{\link{opener.21}},
  \code{\link{opener.2C}}, \code{\link{opener.2Nresp}},
  \code{\link{opener.2N}}, \code{\link{opener.preempt}},
  \code{\link{opener}}
}
\keyword{opener.1N}
