% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{opener.1Xresp}
\alias{opener.1Xresp}
\title{opener.1Xresp}
\usage{
opener.1Xresp(x, y)
}
\arguments{
\item{x}{= internal representation of a bridge hand}

\item{y}{= opening bid}
}
\description{
opener's subsequent bids after 1M
}
\seealso{
Other opener: \code{\link{opener.1Nresp}},
  \code{\link{opener.1N}}, \code{\link{opener.1major}},
  \code{\link{opener.1minor}}, \code{\link{opener.21}},
  \code{\link{opener.2C}}, \code{\link{opener.2Nresp}},
  \code{\link{opener.2N}}, \code{\link{opener.preempt}},
  \code{\link{opener}}
}
\keyword{opener.1Xresp}
