% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{opener.21}
\alias{opener.21}
\title{opener.21}
\usage{
opener.21(x, y)
}
\arguments{
\item{x}{= internal representation of a bridge hand}

\item{y}{= opening bid}
}
\description{
opener's subsequent bids
}
\seealso{
Other opener: \code{\link{opener.1Nresp}},
  \code{\link{opener.1N}}, \code{\link{opener.1Xresp}},
  \code{\link{opener.1major}}, \code{\link{opener.1minor}},
  \code{\link{opener.2C}}, \code{\link{opener.2Nresp}},
  \code{\link{opener.2N}}, \code{\link{opener.preempt}},
  \code{\link{opener}}
}
\keyword{opener.21}
