% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{opener.2C}
\alias{opener.2C}
\title{opener.2C}
\usage{
opener.2C(x)
}
\arguments{
\item{x}{= internal representation of a bridge hand}
}
\description{
opener's subsequent bids after 2C
}
\seealso{
Other opener: \code{\link{opener.1Nresp}},
  \code{\link{opener.1N}}, \code{\link{opener.1Xresp}},
  \code{\link{opener.1major}}, \code{\link{opener.1minor}},
  \code{\link{opener.21}}, \code{\link{opener.2Nresp}},
  \code{\link{opener.2N}}, \code{\link{opener.preempt}},
  \code{\link{opener}}
}
\keyword{opener.2C}
