% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{opener.2N}
\alias{opener.2N}
\title{opener.2N}
\usage{
opener.2N(x)
}
\arguments{
\item{x}{= internal representation of a bridge hand}
}
\description{
opener's subsequent bids after 2N
}
\seealso{
Other opener: \code{\link{opener.1Nresp}},
  \code{\link{opener.1N}}, \code{\link{opener.1Xresp}},
  \code{\link{opener.1minorajor}},
  \code{\link{opener.1minor}}, \code{\link{opener.21}},
  \code{\link{opener.2C}}, \code{\link{opener.2Nresp}},
  \code{\link{opener.preempt}}, \code{\link{opener}}
}
\keyword{opener.2N}
