% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{opener.2Nresp}
\alias{opener.2Nresp}
\title{opener.2Nresp}
\usage{
opener.2Nresp(x, y)
}
\arguments{
\item{x}{= internal representation of a bridge hand}

\item{y}{= opening bid}
}
\description{
opener's subsequent bids after 1M
}
\seealso{
Other opener: \code{\link{opener.1Nresp}},
  \code{\link{opener.1N}}, \code{\link{opener.1Xresp}},
  \code{\link{opener.1minorajor}},
  \code{\link{opener.1minor}}, \code{\link{opener.21}},
  \code{\link{opener.2C}}, \code{\link{opener.2N}},
  \code{\link{opener.preempt}}, \code{\link{opener}}
}
\keyword{opener.2Nresp}
