% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{opener}
\alias{opener}
\title{opener}
\usage{
opener(x)
}
\arguments{
\item{x}{= internal representation of a bridge hand}
}
\description{
responder's response to 1N
}
\seealso{
Other opener: \code{\link{opener.1Nresp}},
  \code{\link{opener.1N}}, \code{\link{opener.1Xresp}},
  \code{\link{opener.1minorajor}},
  \code{\link{opener.1minor}}, \code{\link{opener.21}},
  \code{\link{opener.2C}}, \code{\link{opener.2Nresp}},
  \code{\link{opener.2N}}, \code{\link{opener.preempt}}
}
\keyword{opener}
