% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{resp.1major}
\alias{resp.1major}
\title{resp.1major}
\usage{
resp.1major(x, y)
}
\arguments{
\item{x}{= internal representation of a bridge hand}

\item{y}{= string (opener's bid)}
}
\description{
responder's response to 1M
}
\seealso{
Other resp: \code{\link{resp.1N}},
  \code{\link{resp.1minor}},
  \code{\link{resp.stayman.wrapper}},
  \code{\link{resp.stayman}}, \code{\link{resp}}
}
\keyword{resp.1major}
