% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{resp.1N}
\alias{resp.1N}
\title{resp.1N}
\usage{
resp.1N(x)
}
\arguments{
\item{x}{= internal representation of a bridge hand}
}
\description{
responder's response to 1N
}
\seealso{
Other resp: \code{\link{resp.1major}},
  \code{\link{resp.1minor}},
  \code{\link{resp.stayman.wrapper}},
  \code{\link{resp.stayman}}, \code{\link{resp}}
}
\keyword{resp.1N}
