% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{resp}
\alias{resp}
\title{resp}
\usage{
resp(x, y)
}
\arguments{
\item{x}{= internal representation of a bridge hand}

\item{y}{= string (opener's bid)}
}
\description{
responder's bid
}
\seealso{
Other resp: \code{\link{resp.1N}},
  \code{\link{resp.1minorajor}}, \code{\link{resp.1minor}},
  \code{\link{resp.stayman.wrapper}},
  \code{\link{resp.stayman}}
}
\keyword{resp}
