% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{resp.stayman.wrapper}
\alias{resp.stayman.wrapper}
\title{resp.stayman.wrapper}
\usage{
resp.stayman.wrapper(x)
}
\arguments{
\item{x}{= internal representation of a bridge hand}
}
\description{
responder's response to 1N
}
\seealso{
Other resp: \code{\link{resp.1N}},
  \code{\link{resp.1minorajor}}, \code{\link{resp.1minor}},
  \code{\link{resp.stayman}}, \code{\link{resp}}
}
\keyword{resp.stayman.wrapper}
