% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{responder.1minor}
\alias{responder.1minor}
\title{responder.1minor}
\usage{
responder.1minor(x, y)
}
\arguments{
\item{x}{= internal representation of a bridge hand}

\item{y}{= string (opener's bid)}
}
\description{
responder's response to 1m
}
\seealso{
Other responder: \code{\link{responder.1Nresp}},
  \code{\link{responder.1N}},
  \code{\link{responder.1Xresp}},
  \code{\link{responder.1major}},
  \code{\link{responder.21}}, \code{\link{responder.2C}},
  \code{\link{responder.2Nresp}},
  \code{\link{responder.2N}},
  \code{\link{responder.4thSuit}},
  \code{\link{responder.preempt}}, \code{\link{responder}}
}
\keyword{responder.1minor}
