% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{responder.1N}
\alias{responder.1N}
\title{responder.1N}
\usage{
responder.1N(x)
}
\arguments{
\item{x}{= internal representation of a bridge hand}
}
\description{
responder's response to 1N
}
\seealso{
Other responder: \code{\link{responder.1Nresp}},
  \code{\link{responder.1Xresp}},
  \code{\link{responder.1major}},
  \code{\link{responder.1minor}},
  \code{\link{responder.21}}, \code{\link{responder.2C}},
  \code{\link{responder.2Nresp}},
  \code{\link{responder.2N}},
  \code{\link{responder.4thSuit}},
  \code{\link{responder.preempt}}, \code{\link{responder}}
}
\keyword{responder.1N}
