% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{responder.21}
\alias{responder.21}
\title{responder.21}
\usage{
responder.21(x, y, h)
}
\arguments{
\item{x}{= internal representation of a bridge hand}

\item{y}{= opener's initial bid}

\item{h}{= responder's initial response}
}
\description{
opener's subsequent bids after 1M
}
\seealso{
Other responder: \code{\link{responder.1Nresp}},
  \code{\link{responder.1N}},
  \code{\link{responder.1Xresp}},
  \code{\link{responder.1major}},
  \code{\link{responder.1minor}},
  \code{\link{responder.2C}},
  \code{\link{responder.2Nresp}},
  \code{\link{responder.2N}},
  \code{\link{responder.4thSuit}},
  \code{\link{responder.preempt}}, \code{\link{responder}}
}
\keyword{responder.21}
