% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{responder.2N}
\alias{responder.2N}
\title{responder.2N}
\usage{
responder.2N(x, y = 20)
}
\arguments{
\item{x}{= internal representation of a bridge hand}

\item{y}{= points promised by opener}
}
\description{
responder's response to 2N
}
\seealso{
Other responder: \code{\link{responder.1Nresp}},
  \code{\link{responder.1N}},
  \code{\link{responder.1Xresp}},
  \code{\link{responder.1major}},
  \code{\link{responder.1minor}},
  \code{\link{responder.21}}, \code{\link{responder.2C}},
  \code{\link{responder.2Nresp}},
  \code{\link{responder.4thSuit}},
  \code{\link{responder.preempt}}, \code{\link{responder}}
}
\keyword{responder.2N}
