% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{responder.2Nresp}
\alias{responder.2Nresp}
\title{responder.2Nresp}
\usage{
responder.2Nresp(x, y)
}
\arguments{
\item{x}{= internal representation of a bridge hand}

\item{y}{= string (opener's bid)}
}
\description{
responder's response to <y> 2N
}
\seealso{
Other responder: \code{\link{responder.1Nresp}},
  \code{\link{responder.1N}},
  \code{\link{responder.1Xresp}},
  \code{\link{responder.1major}},
  \code{\link{responder.1minor}},
  \code{\link{responder.21}}, \code{\link{responder.2C}},
  \code{\link{responder.2N}},
  \code{\link{responder.4thSuit}},
  \code{\link{responder.preempt}}, \code{\link{responder}}
}
\keyword{responder.2Nresp}
