% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{responder.4thSuit}
\alias{responder.4thSuit}
\title{responder.4thSuit}
\usage{
responder.4thSuit(bid, n, v, r, u, k)
}
\arguments{
\item{bid}{= auction}

\item{n}{= suit distribution}

\item{v}{= an integer (opener's first suit)}

\item{r}{= an integer (responder's first suit)}

\item{u}{= an integer (opener's second suit)}

\item{k}{= an integer (fourth suit)}
}
\description{
responder's subsequent bids after fourth suit forcing
}
\seealso{
Other responder: \code{\link{responder.1Nresp}},
  \code{\link{responder.1N}},
  \code{\link{responder.1Xresp}},
  \code{\link{responder.1major}},
  \code{\link{responder.1minor}},
  \code{\link{responder.21}}, \code{\link{responder.2C}},
  \code{\link{responder.2Nresp}},
  \code{\link{responder.2N}},
  \code{\link{responder.preempt}}, \code{\link{responder}}
}
\keyword{responder.4thSuit}
