% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{response}
\alias{response}
\title{response}
\usage{
response(x, y, n, w)
}
\arguments{
\item{x}{= list object (bid)}

\item{y}{= suit of partner's response}

\item{n}{= suit of your subsequent response}

\item{w}{= minimum level of response}
}
\description{
response to partner's response
}
\keyword{response}
