% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{response.wrapper}
\alias{response.wrapper}
\title{response.wrapper}
\usage{
response.wrapper(x, y, n, w, h, u, v)
}
\arguments{
\item{x}{= internal representation of a bridge hand}

\item{y}{= suit to respond in}

\item{n}{= detailed suit count}

\item{w}{= T/F depending on whether NT is ok}

\item{h}{= suit of response}

\item{u}{= threshold for each response}

\item{v}{= 1/0 depending on whether meeting threshold confirms fit}
}
\description{
response
}
\keyword{response.wrapper}
