% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{shortness}
\alias{shortness}
\title{shortness}
\usage{
shortness(x, y)
}
\arguments{
\item{x}{= integer vector (detail shape)}

\item{y}{= string (opener's bid)}
}
\description{
shortness outside partner's suit
}
\keyword{shortness}
