% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{stayman}
\alias{stayman}
\title{stayman}
\usage{
stayman(x, y = 2)
}
\arguments{
\item{x}{= internal representation of a bridge hand}

\item{y}{= level of response}
}
\description{
Opener's response to Stayman
}
\keyword{stayman}
