% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{stopper}
\alias{stopper}
\title{stopper}
\usage{
stopper(x)
}
\arguments{
\item{x}{= integer vector (internal representation of cards in a suit)}
}
\description{
T/F depending on whether suit is stopped
}
\keyword{stopper}
