% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{txt.left}
\alias{txt.left}
\title{txt.left}
\usage{
txt.left(x, y)
}
\arguments{
\item{x}{= a string vector}

\item{y}{= a positive integer}
}
\description{
Returns the left <y> characters
}
\seealso{
Other txt: \code{\link{txt.regexp}},
  \code{\link{txt.right}}, \code{\link{txt.trim.left}},
  \code{\link{txt.trim.right}}, \code{\link{txt.trim}}
}
\keyword{txt.left}
