% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{txt.regexp}
\alias{txt.regexp}
\title{txt.regexp}
\usage{
txt.regexp(x)
}
\arguments{
\item{x}{= a string}
}
\description{
converts <x> to a regular expression by padding certain characters with \\\\
}
\seealso{
Other txt: \code{\link{txt.left}}, \code{\link{txt.right}},
  \code{\link{txt.trim.left}},
  \code{\link{txt.trim.right}}, \code{\link{txt.trim}}
}
\keyword{txt.regexp}
