% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{txt.right}
\alias{txt.right}
\title{txt.right}
\usage{
txt.right(x, y)
}
\arguments{
\item{x}{= a string vector}

\item{y}{= a positive integer}
}
\description{
Returns the right <y> characters
}
\seealso{
Other txt: \code{\link{txt.left}},
  \code{\link{txt.regexp}}, \code{\link{txt.trim.left}},
  \code{\link{txt.trim.right}}, \code{\link{txt.trim}}
}
\keyword{txt.right}
