% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duplicate.r
\name{txt.trim.right}
\alias{txt.trim.right}
\title{txt.trim.right}
\usage{
txt.trim.right(x, y)
}
\arguments{
\item{x}{= a string vector}

\item{y}{= a string}
}
\description{
trims off trailing occurences of <y>
}
\seealso{
Other txt: \code{\link{txt.left}},
  \code{\link{txt.regexp}}, \code{\link{txt.right}},
  \code{\link{txt.trim.left}}, \code{\link{txt.trim}}
}
\keyword{txt.trim.right}
